requirements-full.txt
black
cryptography
fonttools
matplotlib
Pillow
python-barcode
py_avataaars
requests
segno
setuptools